backend/requirements.txt
fastapi
uvicorn
sqlmodel
sqlite-utils
python-dotenv
httpx
openai
pinecone
numpy
tweepy
flask
waitress
xdk
